Changed the distance function from int to double because the distance is returned as double not integer

+++ main.ino
@@ -1,7 +1,7 @@
  #include <stdio.h>
  #include <math.h>
 
- int distance(double x, double y){
+ double distance(double x, double y){
     double distance = sqrt(pow(x,2) + pow(y,2));
     return distance;
  }
